Split out several small groupings of tests
As we've discussed before `test_intercept.js` is particularly unmanageable.

Rather than take a complete pass through it, I looked for topics where there are at least a few tests, and there is either some affinity among them (e.g. streaming, http.ClientRequest) making them easier to understand as a group, or they are testing the same function (e.g. `delay*()`, `query()`).

Looking forward to feedback on this, though also hoping we can merge it quickly so it doesn't interfere or create conflicts with other potential work in progress.
