+++ tests/test_intercept.js
@@ -1,14 +1,10 @@
 'use strict'
 
-const fs = require('fs')
 const path = require('path')
 const nock = require('../.')
 const url = require('url')
 const http = require('http')
 const https = require('https')
-const util = require('util')
-const events = require('events')
-const stream = require('stream')
 const { test } = require('tap')
 const mikealRequest = require('request')
 const superagent = require('superagent')
@@ -19,8 +15,6 @@ const hyperquest = require('hyperquest')
 const async = require('async')
 const got = require('got')
 const lolex = require('lolex')
-
-const ssl = require('./ssl')
 
 const textFile = path.join(__dirname, '..', 'assets', 'reply_file_1.txt')
 const binaryFile = path.join(__dirname, '..', 'assets', 'reply_file_2.txt.gz')
@@ -56,37 +50,6 @@ test('double activation throws exception', t => {
   t.true(nock.isActive())
 
   t.end()
-})
-
-test('with allowUnmocked, mocked request still works', async t => {
-  const scope = nock('http://example.com', { allowUnmocked: true })
-    .post('/post')
-    .reply(200, '99problems')
-
-  const { body, statusCode } = await got.post('http://example.com/post')
-  t.equal(statusCode, 200)
-  t.equal(body, '99problems')
-
-  scope.done()
-})
-
-test('allow unmocked works after one interceptor is removed', async t => {
-  const server = http.createServer((request, response) => {
-    response.write('live')
-    response.end()
-  })
-  t.once('end', () => server.close())
-
-  await new Promise(resolve => server.listen(resolve))
-
-  const url = `http://localhost:${server.address().port}`
-
-  nock(url, { allowUnmocked: true })
-    .get('/')
-    .reply(200, 'Mocked')
-
-  t.equal((await got(url)).body, 'Mocked')
-  t.equal((await got(url)).body, 'live')
 })
 
 test("when request's content-type is json: reply callback's requestBody should automatically parse to JSON", async t => {
@@ -819,39 +782,6 @@ test('reply with file', async t => {
   scope.done()
 })
 
-// TODO convert to async / got.
-test('reply with file and pipe response', t => {
-  nock('http://example.test')
-    .get('/')
-    .replyWithFile(200, textFile)
-
-  http.get(
-    {
-      host: 'example.test',
-      path: '/',
-      port: 80,
-    },
-    res => {
-      let str = ''
-      const fakeStream = new (require('stream')).Stream()
-      fakeStream.writable = true
-
-      fakeStream.write = d => {
-        str += d
-      }
-
-      fakeStream.end = () => {
-        t.equal(str, 'Hello from the file!', 'response should match')
-        t.end()
-      }
-
-      res.pipe(fakeStream)
-      res.setEncoding('utf8')
-      t.equal(res.statusCode, 200)
-    }
-  )
-})
-
 test('reply with file with headers', async t => {
   const scope = nock('http://example.test')
     .get('/')
@@ -1056,154 +986,6 @@ test('pause response before data', t => {
   req.end()
 })
 
-// TODO Convert to async / got.
-test('pause response after data', t => {
-  const response = new stream.PassThrough()
-  const scope = nock('http://example.test')
-    .get('/')
-    // Node does not pause the 'end' event so we need to use a stream to simulate
-    // multiple 'data' events.
-    .reply(200, response)
-
-  http.get(
-    {
-      host: 'example.test',
-      path: '/',
-    },
-    res => {
-      let waited = false
-      setTimeout(() => {
-        waited = true
-        res.resume()
-      }, 500)
-
-      res.on('data', data => res.pause())
-
-      res.on('end', () => {
-        t.true(waited)
-        scope.done()
-        t.end()
-      })
-    }
-  )
-
-  // Manually simulate multiple 'data' events.
-  response.emit('data', 'one')
-  setTimeout(() => {
-    response.emit('data', 'two')
-    response.end()
-  }, 0)
-})
-
-// TODO Convert to async / got.
-test('response pipe', t => {
-  const dest = (() => {
-    function Constructor() {
-      events.EventEmitter.call(this)
-
-      this.buffer = Buffer.alloc(0)
-      this.writable = true
-    }
-
-    util.inherits(Constructor, events.EventEmitter)
-
-    Constructor.prototype.end = function() {
-      this.emit('end')
-    }
-
-    Constructor.prototype.write = function(chunk) {
-      const buf = Buffer.alloc(this.buffer.length + chunk.length)
-
-      this.buffer.copy(buf)
-      chunk.copy(buf, this.buffer.length)
-
-      this.buffer = buf
-
-      return true
-    }
-
-    return new Constructor()
-  })()
-
-  const scope = nock('http://example.test')
-    .get('/')
-    .reply(200, 'nobody')
-
-  http.get(
-    {
-      host: 'example.test',
-      path: '/',
-    },
-    res => {
-      dest.on('pipe', () => t.pass('should emit "pipe" event'))
-
-      dest.on('end', () => {
-        scope.done()
-        t.equal(dest.buffer.toString(), 'nobody')
-        t.end()
-      })
-
-      res.pipe(dest)
-    }
-  )
-})
-
-// TODO Convert to async / got.
-test('response pipe without implicit end', t => {
-  const dest = (() => {
-    function Constructor() {
-      events.EventEmitter.call(this)
-
-      this.buffer = Buffer.alloc(0)
-      this.writable = true
-    }
-
-    util.inherits(Constructor, events.EventEmitter)
-
-    Constructor.prototype.end = function() {
-      this.emit('end')
-    }
-
-    Constructor.prototype.write = function(chunk) {
-      const buf = Buffer.alloc(this.buffer.length + chunk.length)
-
-      this.buffer.copy(buf)
-      chunk.copy(buf, this.buffer.length)
-
-      this.buffer = buf
-
-      return true
-    }
-
-    return new Constructor()
-  })()
-
-  const scope = nock('http://example.test')
-    .get('/')
-    .reply(200, 'nobody')
-
-  http.get(
-    {
-      host: 'example.test',
-      path: '/',
-    },
-    res => {
-      dest.on('end', () => t.fail('should not call end implicitly'))
-
-      res.on('end', () => {
-        scope.done()
-        t.pass('should emit end event')
-        t.end()
-      })
-
-      res.pipe(
-        dest,
-        { end: false }
-      )
-    }
-  )
-})
-
 test('chaining API', async t => {
   const scope = nock('http://example.test')
     .get('/one')
@@ -1383,70 +1165,6 @@ test('emits error if https route is missing', t => {
   })
 })
 
-// This test seems to need `http`.
-test('can use ClientRequest using GET', t => {
-  let dataCalled = false
-
-  const scope = nock('http://example.test')
-    .get('/dsad')
-    .reply(202, 'HEHE!')
-
-  const req = new http.ClientRequest({
-    host: 'example.test',
-    path: '/dsad',
-  })
-  req.end()
-
-  req.on('response', function(res) {
-    t.equal(res.statusCode, 202)
-    res.on('end', function() {
-      t.ok(dataCalled, 'data event was called')
-      scope.done()
-      t.end()
-    })
-    res.on('data', function(data) {
-      dataCalled = true
-      t.ok(data instanceof Buffer, 'data should be buffer')
-      t.equal(data.toString(), 'HEHE!', 'response should match')
-    })
-  })
-
-  req.end()
-})
-
-// This test seems to need `http`.
-test('can use ClientRequest using POST', t => {
-  let dataCalled = false
-
-  const scope = nock('http://example.test')
-    .post('/posthere/please', 'heyhey this is the body')
-    .reply(201, 'DOOONE!')
-
-  const req = new http.ClientRequest({
-    host: 'example.test',
-    path: '/posthere/please',
-    method: 'POST',
-  })
-  req.write('heyhey this is the body')
-  req.end()
-
-  req.on('response', function(res) {
-    t.equal(res.statusCode, 201)
-    res.on('end', function() {
-      t.ok(dataCalled, 'data event was called')
-      scope.done()
-      t.end()
-    })
-    res.on('data', function(data) {
-      dataCalled = true
-      t.ok(data instanceof Buffer, 'data should be buffer')
-      t.equal(data.toString(), 'DOOONE!', 'response should match')
-    })
-  })
-
-  req.end()
-})
-
 test('scopes are independent', async t => {
   const scope1 = nock('http://example.test')
     .get('/')
@@ -1477,103 +1195,6 @@ test('two scopes with the same request are consumed', async t => {
 
   scope1.done()
   scope2.done()
-})
-
-test('allow unmocked option works', t => {
-  t.plan(7)
-
-  const server = http.createServer((request, response) => {
-    t.pass('server received a request')
-
-    switch (url.parse(request.url).pathname) {
-      case '/':
-        response.writeHead(200)
-        response.write('server served a response')
-        break
-      case '/not/available':
-        response.writeHead(404)
-        break
-      case '/abc':
-        response.writeHead(200)
-        response.write('server served a response')
-        break
-    }
-
-    response.end()
-  })
-
-  server.listen(() => {
-    const scope = nock(`http://localhost:${server.address().port}`, {
-      allowUnmocked: true,
-    })
-      .get('/abc')
-      .reply(304, 'served from our mock')
-      .get('/wont/get/here')
-      .reply(304, 'served from our mock')
-
-    function secondIsDone() {
-      t.ok(!scope.isDone())
-
-      http
-        .request(
-          {
-            host: 'localhost',
-            path: '/',
-            port: server.address().port,
-          },
-          response => {
-            response.destroy()
-
-            t.assert(response.statusCode == 200, 'Do not intercept /')
-
-            server.close(t.end)
-          }
-        )
-        .end()
-    }
-
-    function firstIsDone() {
-      t.ok(!scope.isDone())
-
-      http
-        .request(
-          {
-            host: 'localhost',
-            path: '/not/available',
-            port: server.address().port,
-          },
-          response => {
-            t.assert(
-              response.statusCode === 404,
-              'Server says it does not exist'
-            )
-
-            response.on('data', function() {})
-            response.on('end', secondIsDone)
-          }
-        )
-        .end()
-    }
-
-    const request = http.request(
-      {
-        host: 'localhost',
-        path: '/abc',
-        port: server.address().port,
-      },
-      response => {
-        t.assert(response.statusCode == 304, 'Intercept /abc')
-
-        response.on('end', firstIsDone)
-        // Streams start in 'paused' mode and must be started.
-        // See https://nodejs.org/api/stream.html#stream_class_stream_readable
-        response.resume()
-      }
-    )
-
-    request.on('error', t.error)
-    request.end()
-  })
 })
 
 test('default reply headers work', t => {
@@ -1789,133 +1410,6 @@ test('pending mocks works', t => {
   })
 })
 
-test("pending mocks doesn't include optional mocks", t => {
-  nock('http://example.test')
-    .get('/nonexistent')
-    .optionally()
-    .reply(200)
-
-  t.deepEqual(nock.pendingMocks(), [])
-  t.end()
-})
-
-test('calling optionally(true) on a mock makes it optional', t => {
-  nock('http://example.test')
-    .get('/nonexistent')
-    .optionally(true)
-    .reply(200)
-
-  t.deepEqual(nock.pendingMocks(), [])
-  t.end()
-})
-
-test('calling optionally(false) on a mock leaves it as required', t => {
-  nock('http://example.test')
-    .get('/nonexistent')
-    .optionally(false)
-    .reply(200)
-
-  t.notEqual(nock.pendingMocks(), [])
-  nock.cleanAll()
-  t.end()
-})
-
-test('optional mocks are still functional', t => {
-  nock('http://example.test')
-    .get('/abc')
-    .optionally()
-    .reply(200)
-
-  http.get({ host: 'example.test', path: '/abc' }, function(res) {
-    t.assert(res.statusCode === 200, 'should still mock requests')
-    t.deepEqual(nock.pendingMocks(), [])
-    t.end()
-  })
-})
-
-test('isDone is true with optional mocks outstanding', t => {
-  const scope = nock('http://example.test')
-    .get('/abc')
-    .optionally()
-    .reply(200)
-
-  t.ok(scope.isDone())
-  t.end()
-})
-
-test('optional but persisted mocks persist, but never appear as pending', async t => {
-  const scope = nock('http://example.test')
-    .get('/')
-    .optionally()
-    .reply(200)
-    .persist()
-
-  t.deepEqual(nock.pendingMocks(), [])
-
-  const response1 = await got('http://example.test/')
-  t.equal(response1.statusCode, 200)
-
-  t.deepEqual(nock.pendingMocks(), [])
-
-  const response2 = await got('http://example.test/')
-  t.equal(response2.statusCode, 200)
-  t.deepEqual(nock.pendingMocks(), [])
-
-  scope.done()
-})
-
-test('optional repeated mocks execute repeatedly, but never appear as pending', t => {
-  nock('http://example.test')
-    .get('/456')
-    .optionally()
-    .times(2)
-    .reply(200)
-
-  t.deepEqual(nock.pendingMocks(), [])
-  http.get({ host: 'example.test', path: '/456' }, function(res) {
-    t.assert(res.statusCode === 200, 'should mock first request')
-    t.deepEqual(nock.pendingMocks(), [])
-
-    http.get({ host: 'example.test', path: '/456' }, function(res) {
-      t.assert(res.statusCode === 200, 'should mock second request')
-      t.deepEqual(nock.pendingMocks(), [])
-      t.end()
-    })
-  })
-})
-
-test("activeMocks returns optional mocks only before they're completed", t => {
-  nock.cleanAll()
-  nock('http://example.test')
-    .get('/optional')
-    .optionally()
-    .reply(200)
-
-  t.deepEqual(nock.activeMocks(), ['GET http://example.test:80/optional'])
-  http.get({ host: 'example.test', path: '/optional' }, function(res) {
-    t.deepEqual(nock.activeMocks(), [])
-    t.end()
-  })
-})
-
-test('activeMocks always returns persisted mocks', async t => {
-  nock.cleanAll()
-
-  const scope = nock('http://example.test')
-    .get('/persisted')
-    .reply(200)
-    .persist()
-
-  t.deepEqual(nock.activeMocks(), ['GET http://example.test:80/persisted'])
-
-  await got('http://example.test/persisted')
-
-  t.deepEqual(nock.activeMocks(), ['GET http://example.test:80/persisted'])
-
-  scope.done()
-  nock.cleanAll()
-})
-
 test('activeMocks returns incomplete mocks', t => {
   nock.cleanAll()
   nock('http://example.test')
@@ -2125,66 +1619,6 @@ test('request has path', t => {
   req.end()
 })
 
-test('persists interceptors', async t => {
-  const scope = nock('http://example.test')
-    .persist()
-    .get('/')
-    .reply(200, 'Persisting all the way')
-
-  t.false(scope.isDone())
-
-  await got('http://example.test/')
-
-  t.true(scope.isDone())
-
-  await got('http://example.test/')
-
-  t.true(scope.isDone())
-
-  nock.cleanAll()
-})
-
-test('Persisted interceptors are in pendingMocks initially', async t => {
-  const scope = nock('http://example.test')
-    .get('/abc')
-    .reply(200, 'Persisted reply')
-    .persist()
-
-  t.deepEqual(scope.pendingMocks(), ['GET http://example.test:80/abc'])
-
-  nock.cleanAll()
-})
-
-test('Persisted interceptors are not in pendingMocks after the first request', async t => {
-  const scope = nock('http://example.test')
-    .get('/def')
-    .reply(200, 'Persisted reply')
-    .persist()
-
-  await got('http://example.test/def')
-
-  t.deepEqual(scope.pendingMocks(), [])
-
-  nock.cleanAll()
-})
-
-test('persist reply with file', async t => {
-  nock('http://example.test')
-    .persist()
-    .get('/')
-    .replyWithFile(200, textFile)
-    .get('/test')
-    .reply(200, 'Yay!')
-
-  for (let i = 0; i < 2; ++i) {
-    const { statusCode, body } = await got('http://example.test/')
-    t.equal(statusCode, 200)
-    t.equal(body, 'Hello from the file!')
-  }
-
-  nock.cleanAll()
-})
-
 test('(re-)activate after restore', t => {
   t.plan(7)
 
@@ -2231,138 +1665,6 @@ test('(re-)activate after restore', t => {
           })
         })
       })
-    })
-  })
-})
-
-test('allow unmocked option works with https', t => {
-  t.plan(6)
-
-  function middleware(request, response) {
-    if (request.url === '/does/not/exist') {
-      response.writeHead(404)
-      response.end()
-      return
-    }
-
-    response.writeHead(200)
-    response.end()
-  }
-
-  ssl.startServer(middleware, function(error, server) {
-    t.error(error)
-
-    const { port } = server.address()
-    const requestOptions = {
-      host: 'localhost',
-      port,
-      ca: ssl.ca,
-    }
-
-    const scope = nock(`https://localhost:${port}`, { allowUnmocked: true })
-      .get('/abc')
-      .reply(200, 'Hey!')
-      .get('/wont/get/here')
-      .reply(200, 'Hi!')
-
-    function secondIsDone() {
-      t.ok(!scope.isDone())
-      https
-        .request(Object.assign({ path: '/' }, requestOptions), res => {
-          res.resume()
-          t.ok(true, 'Google replied to /')
-          res.destroy()
-          t.assert(
-            res.statusCode < 400 && res.statusCode >= 200,
-            'GET Google Home page'
-          )
-
-          server.close(t.end)
-        })
-        .end()
-    }
-
-    function firstIsDone() {
-      t.ok(!scope.isDone(), 'scope is not done')
-      https
-        .request(
-          Object.assign({ path: '/does/not/exist' }, requestOptions),
-          res => {
-            t.equal(404, res.statusCode, 'real google response status code')
-            res.on('data', function() {})
-            res.on('end', secondIsDone)
-          }
-        )
-        .end()
-    }
-
-    https
-      .request(Object.assign({ path: '/abc' }, requestOptions), res => {
-        res.on('end', firstIsDone)
-        // Streams start in 'paused' mode and must be started.
-        // See https://nodejs.org/api/stream.html#stream_class_stream_readable
-        res.resume()
-      })
-      .end()
-  })
-})
-
-test('allow unmocked post with json data', t => {
-  t.plan(2)
-  t.once('end', function() {
-    server.close()
-  })
-
-  const server = http.createServer((request, response) => {
-    t.pass('server received a request')
-    response.writeHead(200)
-    response.end()
-  })
-
-  server.listen(() => {
-    nock(`http://localhost:${server.address().port}`, { allowUnmocked: true })
-      .get('/')
-      .reply(200, 'Hey!')
-
-    const options = {
-      method: 'POST',
-      uri: `http://localhost:${server.address().port}`,
-      json: { some: 'data' },
-    }
-
-    mikealRequest(options, function(err, resp, body) {
-      t.equal(200, resp.statusCode)
-      t.end()
-    })
-  })
-})
-
-test('allow unmocked passthrough with mismatched bodies', t => {
-  t.plan(2)
-  t.once('end', function() {
-    server.close()
-  })
-
-  const server = http.createServer((request, response) => {
-    t.pass('server received a request')
-    response.writeHead(200)
-    response.end()
-  })
-
-  server.listen(() => {
-    nock(`http://localhost:${server.address().port}`, { allowUnmocked: true })
-      .post('/post', { some: 'otherdata' })
-      .reply(404, 'Hey!')
-
-    const options = {
-      method: 'POST',
-      uri: `http://localhost:${server.address().port}/post`,
-      json: { some: 'data' },
-    }
-
-    mikealRequest(options, function(err, resp, body) {
-      t.equal(200, resp.statusCode)
-      t.end()
     })
   })
 })
@@ -2664,37 +1966,6 @@ test('superagent posts', t => {
     })
 })
 
-test('response is streams2 compatible', t => {
-  const responseText = 'streams2 streams2 streams2'
-  nock('http://example.test')
-    .get('/somepath')
-    .reply(200, responseText)
-
-  http
-    .request(
-      {
-        host: 'example.test',
-        path: '/somepath',
-      },
-      function(res) {
-        res.setEncoding('utf8')
-
-        let body = ''
-
-        res.on('readable', function() {
-          let buf
-          while ((buf = res.read())) body += buf
-        })
-
-        res.once('end', function() {
-          t.equal(body, responseText)
-          t.end()
-        })
-      }
-    )
-    .end()
-})
-
 test('response is an http.IncomingMessage instance', t => {
   const responseText = 'incoming message!'
   nock('http://example.test')
@@ -2714,364 +1985,6 @@ test('response is an http.IncomingMessage instance', t => {
       }
     )
     .end()
-})
-
-function checkDuration(t, ms) {
-  // Do not write new tests using this function. Write async tests using
-  // `resolvesInAtLeast` instead.
-  const _end = t.end
-  const start = process.hrtime()
-  let ended = false
-  t.end = function() {
-    if (ended) return
-    ended = true
-    const fin = process.hrtime(start)
-    const finMs =
-      fin[0] * 1000 + // seconds -> ms
-      fin[1] * 1e-6 // nanoseconds -> ms
-
-    /// innaccurate timers
-    ms = ms * 0.9
-
-    t.ok(
-      finMs >= ms,
-      `Duration of ${Math.round(finMs)}ms should be longer than ${ms}ms`
-    )
-    _end.call(t)
-  }
-}
-
-async function resolvesInAtLeast(t, fn, durationMillis) {
-  const startTime = process.hrtime()
-
-  await fn()
-
-  const [seconds, nanoseconds] = process.hrtime(startTime)
-  const elapsedTimeMillis = seconds * 1000 + nanoseconds * 1e-6
-
-  t.ok(
-    elapsedTimeMillis >= durationMillis,
-    `Duration of ${Math.round(
-      elapsedTimeMillis
-    )} ms should be at least ${durationMillis} ms`
-  )
-}
-
-test('calling delay could cause mikealRequest timeout error', t => {
-  const scope = nock('http://example.test')
-    .get('/')
-    .delay({
-      head: 300,
-    })
-    .reply(200, 'OK')
-
-  mikealRequest(
-    {
-      uri: 'http://example.test',
-      method: 'GET',
-      timeout: 100,
-    },
-    function(err) {
-      scope.done()
-      t.equal(err && err.code, 'ESOCKETTIMEDOUT')
-      t.end()
-    }
-  )
-})
-
-test('Body delay does not have impact on timeout', t => {
-  const scope = nock('http://example.test')
-    .get('/')
-    .delay({
-      head: 300,
-      body: 300,
-    })
-    .reply(200, 'OK')
-
-  mikealRequest(
-    {
-      uri: 'http://example.test',
-      method: 'GET',
-      timeout: 500,
-    },
-    function(err, r, body) {
-      t.equal(err, null)
-      t.equal(body, 'OK')
-      t.equal(r.statusCode, 200)
-      scope.done()
-      t.end()
-    }
-  )
-})
-
-test('calling delay with "body" and "head" delays the response', t => {
-  // Do not base new tests on this one. Write async tests using
-  // `resolvesInAtLeast` instead.
-  checkDuration(t, 600)
-
-  nock('http://example.test')
-    .get('/')
-    .delay({
-      head: 300,
-      body: 300,
-    })
-    .reply(200, 'OK')
-
-  http.get('http://example.test', function(res) {
-    res.once('data', function(data) {
-      t.equal(data.toString(), 'OK')
-      res.once('end', t.end.bind(t))
-    })
-  })
-})
-
-test('calling delay with "body" delays the response body', t => {
-  // Do not base new tests on this one. Write async tests using
-  // `resolvesInAtLeast` instead.
-  checkDuration(t, 100)
-
-  nock('http://example.test')
-    .get('/')
-    .delay({
-      body: 100,
-    })
-    .reply(200, 'OK')
-
-  http.get('http://example.test', function(res) {
-    res.setEncoding('utf8')
-
-    let body = ''
-
-    res.on('data', function(chunk) {
-      body += chunk
-    })
-
-    res.once('end', function() {
-      t.equal(body, 'OK')
-      t.end()
-    })
-  })
-})
-
-test('calling delayBody delays the response', async t => {
-  const scope = nock('http://example.test')
-    .get('/')
-    .delayBody(100)
-    .reply(200, 'OK')
-
-  await resolvesInAtLeast(
-    t,
-    async () => {
-      const { body } = await got('http://example.test')
-      t.equal(body, 'OK')
-    },
-    100
-  )
-
-  scope.done()
-})
-
-test('delayBody works with a stream', async t => {
-  const scope = nock('http://example.test')
-    .get('/')
-    .delayBody(100)
-    .reply(200, (uri, requestBody) =>
-      fs.createReadStream(textFile, { encoding: 'utf8' })
-    )
-
-  await resolvesInAtLeast(
-    t,
-    async () => {
-      const { body } = await got('http://example.test')
-      t.equal(body, fs.readFileSync(textFile, { encoding: 'utf8' }))
-    },
-    100
-  )
-
-  scope.done()
-})
-
-test('delayBody works with a stream of binary buffers', async t => {
-  const scope = nock('http://example.com')
-    .get('/')
-    .delayBody(100)
-    // No encoding specified, which causes the file to be streamed using
-    // buffers instead of strings.
-    .reply(200, (uri, requestBody) => fs.createReadStream(textFile))
-
-  await resolvesInAtLeast(
-    t,
-    async () => {
-      const { body } = await got('http://example.com/')
-      t.equal(body, fs.readFileSync(textFile, { encoding: 'utf8' }))
-    },
-    100
-  )
-
-  scope.done()
-})
-
-test('delayBody works with a delayed stream', async t => {
-  const passthrough = new stream.Transform({
-    transform(chunk, encoding, callback) {
-      this.push(chunk.toString())
-      callback()
-    },
-  })
-
-  const scope = nock('http://example.com')
-    .get('/')
-    .delayBody(100)
-    .reply(200, (uri, requestBody) => passthrough)
-
-  setTimeout(() => fs.createReadStream(textFile).pipe(passthrough), 125)
-
-  const { body } = await got('http://example.com/')
-  t.equal(body, fs.readFileSync(textFile, { encoding: 'utf8' }))
-
-  scope.done()
-})
-
-test('calling delay delays the response', t => {
-  // Do not base new tests on this one. Write async tests using
-  // `resolvesInAtLeast` instead.
-  checkDuration(t, 100)
-
-  nock('http://example.test')
-    .get('/')
-    .delay(100)
-    .reply(200, 'OK')
-
-  http.get('http://example.test', function(res) {
-    res.setEncoding('utf8')
-
-    let body = ''
-
-    res.on('data', function(chunk) {
-      body += chunk
-    })
-
-    res.once('end', function() {
-      t.equal(body, 'OK')
-      t.end()
-    })
-  })
-})
-
-test('using reply callback with delay provides proper arguments', t => {
-  nock('http://localhost')
-    .get('/')
-    .delay(100)
-    .reply(200, function(path, requestBody) {
-      t.equal(path, '/', 'path arg should be set')
-      t.equal(requestBody, 'OK', 'requestBody arg should be set')
-      t.end()
-    })
-
-  http.request('http://localhost/', function() {}).end('OK')
-})
-
-test('using reply callback with delay can reply JSON', t => {
-  nock('http://example.test')
-    .get('/')
-    .delay(100)
-    .reply(200, function(path, requestBody) {
-      return { a: 1 }
-    })
-
-  mikealRequest.get(
-    {
-      url: 'http://example.test/',
-      json: true,
-    },
-    function(err, res, body) {
-      t.equals(res.headers['content-type'], 'application/json')
-      t.deepEqual(body, { a: 1 })
-      t.end()
-    }
-  )
-})
-
-test('delay works with replyWithFile', t => {
-  // Do not base new tests on this one. Write async tests using
-  // `resolvesInAtLeast` instead.
-  checkDuration(t, 100)
-
-  nock('http://localhost')
-    .get('/')
-    .delay(100)
-    .replyWithFile(200, `${__dirname}/../assets/reply_file_1.txt`)
-
-  http
-    .request('http://localhost/', function(res) {
-      res.setEncoding('utf8')
-
-      let body = ''
-
-      res.on('data', function(chunk) {
-        body += chunk
-      })
-
-      res.once('end', function() {
-        t.equal(
-          body,
-          'Hello from the file!',
-          'the body should eql the text from the file'
-        )
-        t.end()
-      })
-    })
-    .end('OK')
-})
-
-test('delay works with when you return a generic stream from the reply callback', t => {
-  // Do not base new tests on this one. Write async tests using
-  // `resolvesInAtLeast` instead.
-  checkDuration(t, 100)
-
-  nock('http://localhost')
-    .get('/')
-    .delay(100)
-    .reply(200, function(path, reqBody) {
-      return fs.createReadStream(`${__dirname}/../assets/reply_file_1.txt`)
-    })
-
-  http
-    .request('http://localhost/', function(res) {
-      res.setEncoding('utf8')
-
-      let body = ''
-
-      res.on('data', function(chunk) {
-        body += chunk
-      })
-
-      res.once('end', function() {
-        t.equal(
-          body,
-          'Hello from the file!',
-          'the body should eql the text from the file'
-        )
-        t.end()
-      })
-    })
-    .end('OK')
-})
-
-test('delay with replyWithError: response is delayed', async t => {
-  nock('http://example.test')
-    .get('/')
-    .delay(100)
-    .replyWithError('this is an error message')
-
-  await resolvesInAtLeast(
-    t,
-    async () =>
-      t.rejects(() => got('http://example.test'), {
-        message: 'this is an error message',
-      }),
-    100
-  )
 })
 
 test('write callback called', t => {
@@ -3171,152 +2084,6 @@ test('finish event fired before end event (bug-139)', t => {
   })
 
   req.end('mamma mia')
-})
-
-test(
-  'when a stream is used for the response body, it will not be read until after the response event',
-  { skip: !stream.Readable },
-  t => {
-    let responseEvent = false
-    const text = 'Hello World\n'
-
-    function SimpleStream(opt) {
-      stream.Readable.call(this, opt)
-    }
-    util.inherits(SimpleStream, stream.Readable)
-    SimpleStream.prototype._read = function() {
-      t.ok(responseEvent)
-      this.push(text)
-      this.push(null)
-    }
-
-    nock('http://localhost')
-      .get('/')
-      .reply(200, function(path, reqBody) {
-        return new SimpleStream()
-      })
-
-    http.get('http://localhost/', function(res) {
-      responseEvent = true
-      res.setEncoding('utf8')
-
-      let body = ''
-
-      res.on('data', function(chunk) {
-        body += chunk
-      })
-
-      res.once('end', function() {
-        t.equal(body, text)
-        t.end()
-      })
-    })
-  }
-)
-
-test('calling delayConnection delays the connection', t => {
-  // Do not base new tests on this one. Write async tests using
-  // `resolvesInAtLeast` instead.
-  checkDuration(t, 100)
-
-  nock('http://example.test')
-    .get('/')
-    .delayConnection(100)
-    .reply(200, 'OK')
-
-  http.get('http://example.test', function(res) {
-    res.setEncoding('utf8')
-
-    let body = ''
-
-    res.on('data', function(chunk) {
-      body += chunk
-    })
-
-    res.once('end', function() {
-      t.equal(body, 'OK')
-      t.end()
-    })
-  })
-})
-
-test('using reply callback with delayConnection provides proper arguments', t => {
-  nock('http://localhost')
-    .get('/')
-    .delayConnection(100)
-    .reply(200, function(path, requestBody) {
-      t.equal(path, '/', 'path arg should be set')
-      t.equal(requestBody, 'OK', 'requestBody arg should be set')
-      t.end()
-    })
-
-  http.request('http://localhost/', function() {}).end('OK')
-})
-
-test('delayConnection works with replyWithFile', t => {
-  // Do not base new tests on this one. Write async tests using
-  // `resolvesInAtLeast` instead.
-  checkDuration(t, 100)
-
-  nock('http://localhost')
-    .get('/')
-    .delayConnection(100)
-    .replyWithFile(200, `${__dirname}/../assets/reply_file_1.txt`)
-
-  http
-    .request('http://localhost/', function(res) {
-      res.setEncoding('utf8')
-
-      let body = ''
-
-      res.on('data', function(chunk) {
-        body += chunk
-      })
-
-      res.once('end', function() {
-        t.equal(
-          body,
-          'Hello from the file!',
-          'the body should eql the text from the file'
-        )
-        t.end()
-      })
-    })
-    .end('OK')
-})
-
-test('delayConnection works with when you return a generic stream from the reply callback', t => {
-  // Do not base new tests on this one. Write async tests using
-  // `resolvesInAtLeast` instead.
-  checkDuration(t, 100)
-
-  nock('http://localhost')
-    .get('/')
-    .delayConnection(100)
-    .reply(200, function(path, reqBody) {
-      return fs.createReadStream(`${__dirname}/../assets/reply_file_1.txt`)
-    })
-
-  http
-    .request('http://localhost/', function(res) {
-      res.setEncoding('utf8')
-
-      let body = ''
-
-      res.on('data', function(chunk) {
-        body += chunk
-      })
-
-      res.once('end', function() {
-        t.equal(
-          body,
-          'Hello from the file!',
-          'the body should eql the text from the file'
-        )
-        t.end()
-      })
-    })
-    .end('OK')
 })
 
 test('sending binary and receiving JSON should work ', t => {
@@ -3701,25 +2468,6 @@ test('done fails when specified bad request header is present', t => {
   })
 })
 
-test('mikeal/request with delayConnection and request.timeout', t => {
-  nock('http://example.test')
-    .post('/')
-    .delayConnection(1000)
-    .reply(200, {})
-
-  mikealRequest.post(
-    {
-      url: 'http://example.test',
-      timeout: 10,
-    },
-    function(err) {
-      t.type(err, 'Error')
-      t.equal(err && err.code, 'ESOCKETTIMEDOUT')
-      t.end()
-    }
-  )
-})
-
 test('get correct filtering with scope and request headers filtering', t => {
   const responseText = 'OK!'
   const responseHeaders = { 'Content-Type': 'text/plain' }
@@ -3851,38 +2599,6 @@ test('mikeal/request with strictSSL: true', t => {
       t.end()
     }
   )
-})
-
-test('response readable pull stream works as expected', t => {
-  nock('http://example.test')
-    .get('/ssstream')
-    .reply(200, 'this is the response body yeah')
-
-  const req = http.request(
-    {
-      host: 'example.test',
-      path: '/ssstream',
-      port: 80,
-    },
-    function(res) {
-      let ended = false
-      let responseBody = ''
-      t.equal(res.statusCode, 200)
-      res.on('readable', function() {
-        let chunk
-        while ((chunk = res.read()) !== null) {
-          responseBody += chunk.toString()
-        }
-        if (chunk === null && !ended) {
-          ended = true
-          t.equal(responseBody, 'this is the response body yeah')
-          t.end()
-        }
-      })
-    }
-  )
-
-  req.end()
 })
 
 test('.setNoDelay', t => {
@@ -4134,40 +2850,6 @@ test('match path using regexp', t => {
   })
 })
 
-test('match path using regexp with allowUnmocked', t => {
-  nock('http://example.test', { allowUnmocked: true })
-    .get(/regex$/)
-    .reply(200, 'Match regex')
-
-  mikealRequest.get('http://example.test/resources/regex', function(
-    err,
-    res,
-    body
-  ) {
-    t.type(err, 'null')
-    t.equal(res.statusCode, 200)
-    t.equal(body, 'Match regex')
-    t.end()
-  })
-})
-
-test('match hostname using regexp with allowUnmocked (issue-1076)', t => {
-  nock(/localhost/, { allowUnmocked: true })
-    .get('/no/regex/here')
-    .reply(200, 'Match regex')
-
-  mikealRequest.get('http://localhost:3000/no/regex/here', function(
-    err,
-    res,
-    body
-  ) {
-    t.type(err, 'null')
-    t.equal(res.statusCode, 200)
-    t.equal(body, 'Match regex')
-    t.end()
-  })
-})
-
 test('match path using function', t => {
   const path = '/match/uri/function'
   const options = {
@@ -4300,67 +2982,6 @@ test('you must setup an interceptor for each request', t => {
   })
 })
 
-test('calling socketDelay will emit a timeout', t => {
-  nock('http://example.test')
-    .get('/')
-    .socketDelay(10000)
-    .reply(200, 'OK')
-
-  let timedout = false
-  let ended = false
-
-  const req = http.request('http://example.test', function(res) {
-    res.setEncoding('utf8')
-
-    res.once('end', function() {
-      ended = true
-      if (!timedout) {
-        t.fail('socket did not timeout when idle')
-        t.end()
-      }
-    })
-  })
-
-  req.setTimeout(5000, function() {
-    timedout = true
-    if (!ended) {
-      t.ok(true)
-      t.end()
-    }
-  })
-
-  req.end()
-})
-
-test('calling socketDelay not emit a timeout if not idle for long enough', t => {
-  nock('http://example.test')
-    .get('/')
-    .socketDelay(10000)
-    .reply(200, 'OK')
-
-  const req = http.request('http://example.test', function(res) {
-    res.setEncoding('utf8')
-
-    let body = ''
-
-    res.on('data', function(chunk) {
-      body += chunk
-    })
-
-    res.once('end', function() {
-      t.equal(body, 'OK')
-      t.end()
-    })
-  })
-
-  req.setTimeout(60000, function() {
-    t.fail('socket timed out unexpectedly')
-    t.end()
-  })
-
-  req.end()
-})
-
 test('replyWithError returns an error on request', t => {
   const scope = nock('http://example.test')
     .post('/echo')
@@ -4434,342 +3055,6 @@ test('no content type provided', t => {
     .end('WHAA')
 })
 
-test('query() matches a query string of the same name=value', t => {
-  nock('http://example.test')
-    .get('/')
-    .query({ foo: 'bar' })
-    .reply(200)
-
-  mikealRequest('http://example.test/?foo=bar', function(err, res) {
-    if (err) throw err
-    t.equal(res.statusCode, 200)
-    t.end()
-  })
-})
-
-test('query() matches multiple query strings of the same name=value', t => {
-  nock('http://example.test')
-    .get('/')
-    .query({ foo: 'bar', baz: 'foz' })
-    .reply(200)
-
-  mikealRequest('http://example.test/?foo=bar&baz=foz', function(err, res) {
-    if (err) throw err
-    t.equal(res.statusCode, 200)
-    t.end()
-  })
-})
-
-test('query() matches multiple query strings of the same name=value regardless of order', t => {
-  nock('http://example.test')
-    .get('/')
-    .query({ foo: 'bar', baz: 'foz' })
-    .reply(200)
-
-  mikealRequest('http://example.test/?baz=foz&foo=bar', function(err, res) {
-    if (err) throw err
-    t.equal(res.statusCode, 200)
-    t.end()
-  })
-})
-
-test('query() matches query values regardless of their type of declaration', t => {
-  nock('http://example.test')
-    .get('/')
-    .query({ num: 1, bool: true, empty: null, str: 'fou' })
-    .reply(200)
-
-  mikealRequest('http://example.test/?num=1&bool=true&empty=&str=fou', function(
-    err,
-    res
-  ) {
-    if (err) throw err
-    t.equal(res.statusCode, 200)
-    t.end()
-  })
-})
-
-test("query() doesn't match query values of requests without query string", t => {
-  nock('http://example.test')
-    .get('/')
-    .query({ num: 1, bool: true, empty: null, str: 'fou' })
-    .reply(200, 'scope1')
-
-  nock('http://example.test')
-    .get('/')
-    .reply(200, 'scope2')
-
-  mikealRequest('http://example.test', function(err, res) {
-    if (err) throw err
-    t.equal(res.statusCode, 200)
-    t.equal(res.body, 'scope2')
-    t.end()
-  })
-})
-
-test('query() matches a query string using regexp', t => {
-  nock('http://example.test')
-    .get('/')
-    .query({ foo: /.*/ })
-    .reply(200)
-
-  mikealRequest('http://example.test/?foo=bar', function(err, res) {
-    if (err) throw err
-    t.equal(res.statusCode, 200)
-    t.end()
-  })
-})
-
-test('query() matches a query string that contains special RFC3986 characters', t => {
-  nock('http://example.test')
-    .get('/')
-    .query({ 'foo&bar': 'hello&world' })
-    .reply(200)
-
-  const options = {
-    uri: 'http://example.test',
-    qs: {
-      'foo&bar': 'hello&world',
-    },
-  }
-
-  mikealRequest(options, function(err, res) {
-    if (err) throw err
-    t.equal(res.statusCode, 200)
-    t.end()
-  })
-})
-
-test('query() expects unencoded query params', t => {
-  nock('http://example.test')
-    .get('/')
-    .query({ foo: 'hello%20world' })
-    .reply(200)
-
-  mikealRequest('http://example.test?foo=hello%20world', function(err, res) {
-    t.similar(err.toString(), /Error: Nock: No match for request/)
-    t.end()
-  })
-})
-
-test('query() matches a query string with pre-encoded values', t => {
-  nock('http://example.test', { encodedQueryParams: true })
-    .get('/')
-    .query({ foo: 'hello%20world' })
-    .reply(200)
-
-  mikealRequest('http://example.test?foo=hello%20world', function(err, res) {
-    if (err) throw err
-    t.equal(res.statusCode, 200)
-    t.end()
-  })
-})
-
-test('query() with "true" will allow all query strings to pass', t => {
-  nock('http://example.test')
-    .get('/')
-    .query(true)
-    .reply(200)
-
-  mikealRequest('http://example.test/?foo=bar&a=1&b=2', function(err, res) {
-    if (err) throw err
-    t.equal(res.statusCode, 200)
-    t.end()
-  })
-})
-
-test('query() with "{}" will allow a match against ending in ?', t => {
-  nock('http://example.test')
-    .get('/noquerystring')
-    .query({})
-    .reply(200)
-
-  mikealRequest('http://example.test/noquerystring?', function(err, res) {
-    if (err) throw err
-    t.equal(res.statusCode, 200)
-    t.end()
-  })
-})
-
-test('query() with a function, function called with actual queryObject', t => {
-  let queryObject
-
-  const queryValidator = function(qs) {
-    queryObject = qs
-    return true
-  }
-
-  nock('http://example.test')
-    .get('/')
-    .query(queryValidator)
-    .reply(200)
-
-  mikealRequest('http://example.test/?foo=bar&a=1&b=2', function(err, res) {
-    if (err) throw err
-    t.deepEqual(queryObject, { foo: 'bar', a: '1', b: '2' })
-    t.equal(res.statusCode, 200)
-    t.end()
-  })
-})
-
-test('query() with a function, function return true the query treat as matched', t => {
-  const alwasyTrue = function() {
-    return true
-  }
-
-  nock('http://example.test')
-    .get('/')
-    .query(alwasyTrue)
-    .reply(200)
-
-  mikealRequest('http://example.test/?igore=the&actual=query', function(
-    err,
-    res
-  ) {
-    if (err) throw err
-    t.equal(res.statusCode, 200)
-    t.end()
-  })
-})
-
-test('query() with a function, function return false the query treat as Un-matched', t => {
-  const alwayFalse = function() {
-    return false
-  }
-
-  nock('http://example.test')
-    .get('/')
-    .query(alwayFalse)
-    .reply(200)
-
-  mikealRequest('http://example.test/?i=should&pass=?', function(err, res) {
-    t.equal(
-      err.message.trim(),
-      `Nock: No match for request ${JSON.stringify(
-        {
-          method: 'GET',
-          url: 'http://example.test/?i=should&pass=?',
-          headers: { host: 'example.test' },
-        },
-        null,
-        2
-      )}`
-    )
-    t.end()
-  })
-})
-
-test('query() will not match when a query string does not match name=value', t => {
-  nock('https://example.test')
-    .get('/b')
-    .query({ foo: 'bar' })
-    .reply(200)
-
-  mikealRequest('https://example.test/b?foo=baz', function(err, res) {
-    t.equal(
-      err.message.trim(),
-      `Nock: No match for request ${JSON.stringify(
-        {
-          method: 'GET',
-          url: 'https://example.test/b?foo=baz',
-          headers: { host: 'example.test' },
-        },
-        null,
-        2
-      )}`
-    )
-    t.end()
-  })
-})
-
-test('query() will not match when a query string is present that was not registered', t => {
-  nock('https://example.test')
-    .get('/c')
-    .query({ foo: 'bar' })
-    .reply(200)
-
-  mikealRequest('https://example.test/c?foo=bar&baz=foz', function(err, res) {
-    t.equal(
-      err.message.trim(),
-      `Nock: No match for request ${JSON.stringify(
-        {
-          method: 'GET',
-          url: 'https://example.test/c?foo=bar&baz=foz',
-          headers: { host: 'example.test' },
-        },
-        null,
-        2
-      )}`
-    )
-    t.end()
-  })
-})
-
-test('query() will not match when a query string is malformed', t => {
-  nock('https://example.test')
-    .get('/d')
-    .query({ foo: 'bar' })
-    .reply(200)
-
-  mikealRequest('https://example.test/d?foobar', function(err, res) {
-    t.equal(
-      err.message.trim(),
-      `Nock: No match for request ${JSON.stringify(
-        {
-          method: 'GET',
-          url: 'https://example.test/d?foobar',
-          headers: { host: 'example.test' },
-        },
-        null,
-        2
-      )}`
-    )
-    t.end()
-  })
-})
-
-test('query() will not match when a query string has fewer correct values than expected', t => {
-  nock('http://example.test')
-    .get('/')
-    .query({
-      num: 1,
-      bool: true,
-      empty: null,
-      str: 'fou',
-    })
-    .reply(200)
-
-  mikealRequest('http://example.test/?num=1str=fou', function(err, res) {
-    t.equal(
-      err.message.trim(),
-      `Nock: No match for request ${JSON.stringify(
-        {
-          method: 'GET',
-          url: 'http://example.test/?num=1str=fou',
-          headers: { host: 'example.test' },
-        },
-        null,
-        2
-      )}`
-    )
-    t.end()
-  })
-})
-
-test('query(true) will match when the path has no query', t => {
-  nock('http://example.test')
-    .get('/')
-    .query(true)
-    .reply(200)
-
-  mikealRequest('http://example.test', function(err, res) {
-    t.ok(!err, 'no error')
-    t.ok(res)
-    t.equal(res.statusCode, 200)
-    t.end()
-  })
-})
-
 // https://github.com/nock/nock/issues/835
 test('match domain and path using regexp', t => {
   nock.cleanAll()
@@ -4777,56 +3062,6 @@ test('match domain and path using regexp', t => {
 
   const scope = nock(/example/)
     .get(/img/)
-    .reply(200, imgResponse)
-
-  mikealRequest.get('http://example.test/imghp?hl=en', function(
-    err,
-    res,
-    body
-  ) {
-    scope.done()
-    t.type(err, 'null')
-    t.equal(res.statusCode, 200)
-    t.equal(body, imgResponse)
-    t.end()
-  })
-})
-
-// https://github.com/nock/nock/issues/835
-test('match multiple paths to domain using regexp with allowUnmocked', async t => {
-  nock.cleanAll()
-
-  const server = http.createServer((request, response) => {
-    response.write('live')
-    response.end()
-  })
-  t.once('end', () => server.close())
-  await new Promise(resolve => server.listen(resolve))
-  const url = `http://localhost:${server.address().port}`
-
-  const scope1 = nock(/localhost/, { allowUnmocked: true })
-    .get(/alpha/)
-    .reply(200, 'this is alpha')
-
-  const scope2 = nock(/localhost/, { allowUnmocked: true })
-    .get(/bravo/)
-    .reply(200, 'bravo, bravo!')
-
-  t.equal((await got(`${url}`)).body, 'live')
-  t.equal((await got(`${url}/alphalicious`)).body, 'this is alpha')
-  t.equal((await got(`${url}/bravo-company`)).body, 'bravo, bravo!')
-
-  scope1.done()
-  scope2.done()
-})
-
-test('match domain and path using regexp with query params and allow unmocked', t => {
-  nock.cleanAll()
-  const imgResponse = 'Matched Images Page'
-  const opts = { allowUnmocked: true }
-
-  const scope = nock(/example/, opts)
-    .get(/imghp\?hl=en/)
     .reply(200, imgResponse)
 
   mikealRequest.get('http://example.test/imghp?hl=en', function(
@@ -4896,29 +3131,6 @@ test('multiple interceptors override headers from unrelated request', t => {
   )
 })
 
-// https://github.com/nock/nock/issues/490
-test('match when query is specified with allowUnmocked', async t => {
-  nock.cleanAll()
-
-  const server = http.createServer((request, response) => {
-    response.write('live')
-    response.end()
-  })
-  t.once('end', () => server.close())
-  await new Promise(resolve => server.listen(resolve))
-  const url = `http://localhost:${server.address().port}`
-
-  const scope = nock(url, { allowUnmocked: true })
-    .get('/search')
-    .query({ q: 'cat pictures' })
-    .reply(200, '😻')
-
-  t.equal((await got(url)).body, 'live')
-  t.equal((await got(`${url}/search?q=cat%20pictures`)).body, '😻')
-
-  scope.done()
-})
-
 // https://github.com/nock/nock/issues/1003
 test('correctly parse request without specified path', t => {
   nock.cleanAll()
@@ -4960,40 +3172,6 @@ test('data is sent with flushHeaders', t => {
     .flushHeaders()
 })
 
-test('stop persisting a persistent nock', async t => {
-  nock.cleanAll()
-
-  const scope = nock('http://example.test')
-    .persist(true)
-    .get('/')
-    .reply(200, 'Persisting all the way')
-
-  t.false(scope.isDone())
-
-  await got('http://example.test/')
-
-  t.true(scope.isDone())
-  t.deepEqual(nock.activeMocks(), ['GET http://example.test:80/'])
-
-  scope.persist(false)
-
-  await got('http://example.test/')
-
-  t.equal(nock.activeMocks().length, 0)
-  t.true(scope.isDone())
-
-  await t.rejects(async () => got('http://example.test/'), {
-    message: 'Nock: No match for request',
-  })
-})
-
-test("should throw an error when persist flag isn't a boolean", t => {
-  t.throws(() => nock('http://persist.com').persist('string'), {
-    message: 'Invalid arguments: argument should be a boolean',
-  })
-  t.end()
-})
-
 test('should log matching', async t => {
   const messages = []
 
@@ -5009,51 +3187,6 @@ test('should log matching', async t => {
   ])
 
   scope.done()
-})
-
-test('creating ClientRequest with empty options throws expected error', t => {
-  // Confidence check.
-  t.ok(nock.isActive())
-
-  t.throws(() => new http.ClientRequest(), {
-    message:
-      'Creating a client request with empty `options` is not supported in Nock',
-  })
-
-  t.end()
-})
-
-test('when no interceptors and net connect is allowed, request via ClientRequest goes through', t => {
-  nock.cleanAll()
-  nock.enableNetConnect()
-
-  // Confidence check. We need nock enabled to test the desired path.
-  t.ok(nock.isActive())
-
-  const server = http.createServer((request, response) => {
-    response.writeHead(201)
-    response.end()
-  })
-  t.once('end', () => server.close())
-
-  server.listen(() => {
-    const req = new http.ClientRequest({ port: server.address().port })
-    req.on('response', res => {
-      t.equal(res.statusCode, 201)
-      t.end()
-    })
-    req.end()
-  })
-})
-
-test('when no interceptors and net connect is disallowed, receive via ClientRequest emits the expected error', t => {
-  nock.disableNetConnect()
-  t.once('end', () => nock.enableNetConnect())
-
-  new http.ClientRequest({ port: 12345, path: '/' }).on('error', err => {
-    t.equal(err.message, 'Nock: Disallowed net connect for "localhost:12345/"')
-    t.end()
-  })
 })
 
 test('teardown', t => {
