+++ tests/test_socketdelay.js
@@ -34,3 +34,64 @@ test('socketDelay', function(t) {
     t.end()
   })
 })
+
+test('calling socketDelay will emit a timeout', t => {
+  nock('http://example.test')
+    .get('/')
+    .socketDelay(10000)
+    .reply(200, 'OK')
+
+  let timedout = false
+  let ended = false
+
+  const req = http.request('http://example.test', function(res) {
+    res.setEncoding('utf8')
+
+    res.once('end', function() {
+      ended = true
+      if (!timedout) {
+        t.fail('socket did not timeout when idle')
+        t.end()
+      }
+    })
+  })
+
+  req.setTimeout(5000, function() {
+    timedout = true
+    if (!ended) {
+      t.ok(true)
+      t.end()
+    }
+  })
+
+  req.end()
+})
+
+test('calling socketDelay not emit a timeout if not idle for long enough', t => {
+  nock('http://example.test')
+    .get('/')
+    .socketDelay(10000)
+    .reply(200, 'OK')
+
+  const req = http.request('http://example.test', function(res) {
+    res.setEncoding('utf8')
+
+    let body = ''
+
+    res.on('data', function(chunk) {
+      body += chunk
+    })
+
+    res.once('end', function() {
+      t.equal(body, 'OK')
+      t.end()
+    })
+  })
+
+  req.setTimeout(60000, function() {
+    t.fail('socket timed out unexpectedly')
+    t.end()
+  })
+
+  req.end()
+})
